requirements.txt
astroid==2.12.9
dill==0.3.5.1
gitdb==4.0.9
GitPython==3.1.27
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
py-markdown-table==0.3.3
pylint==2.15.2
smmap==5.0.0
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
wrapt==1.14.1
